pyproject.toml
[tool.poetry]
name = "drf-viewsets-generics-24-1"
version = "0.1.0"
description = ""
authors = ["Vladimir Titov <tvv-vg@yandex.ru>"]
readme = "README.md"
packages = [{include = "drf_viewsets_generics_24"}]

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.6"
djangorestframework = "^3.15.1"
psycopg2-binary = "^2.9.9"
pillow = "^10.3.0"
python-dotenv = "^1.0.1"
django-filter = "^24.2"
djangorestframework-simplejwt = "^5.3.1"


[tool.poetry.group.lint.dependencies]
flake8 = "^7.0.0"
isort = "^5.13.2"
mypy = "^1.10.0"


[tool.poetry.group.test.dependencies]
pytest = "^8.2.2"
pytest-django = "^4.8.0"
coverage = "^7.5.3"

[tool.mypy]
disallow_untyped_defs = true
no_implicit_optional = true
warn_return_any = true
exclude = 'venv'

[tool.isort]
# максимальная длина строки
line_length = 119

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
